+++ Poseiden-skeleton/src/main/java/com/poseidon/api/repositories/RuleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RuleRepository extends JpaRepository<Rule, Integer> {
+public interface RuleRepository extends JpaRepository<Rule, Long> {
 }
